feat: refactor package dependencies fetching and move to new api/packages.ts
