+++ src/components/DependenciesList.tsx
@@ -1,12 +1,14 @@
 import React, { useEffect, useState } from 'react';
 import { Box, Table, Text, Badge } from '@radix-ui/themes';
-import { Dependency } from '../api/types';
-import { fetchPackageDependencies } from '../api/packageApi';
+import { fetchPackageDependencies } from '../api/packages';
+
+import type { Dependency } from '../api/types';
 
 interface DependenciesListProps {
   packageId: string;
   version: string;
 }
+
 
 export const DependenciesList: React.FC<DependenciesListProps> = ({ packageId, version }) => {
   const [dependencies, setDependencies] = useState<Dependency[]>([]);
